Use pkg-config for detecting tinyxml
The detection of tinyxml fails in debian because the library is no
longer installed into /usr/lib, is now in /usr/lib/$(ARCH_TRIPLET), for
instance /usr/lib/x86_64-linux-gnu. The upstream included
Find_tinyxml.cmake detects the Headers correctly, but breaks the build
for failing to locate the libraries.

Tinyxml ships a working pkg-config file, so let's use that one for now.